third_party/WebKit/LayoutTests/platform/win/http/tests/security/mixedContent/insecure-localhost-allowed.https-expected.txt

CONSOLE ERROR: line 18: Mixed Content: The page at 'https://127.0.0.1:8443/security/mixedContent/insecure-localhost-allowed.https.html' was loaded over HTTPS, but requested an insecure resource 'http://example.test:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=*'. This request has been blocked; the content must be served over HTTPS.
This is a testharness.js-based test.
PASS Fetching 'http://127.0.0.1/' should not be blocked. 
PASS Fetching 'http://example.test/' should be blocked. 
Harness: the test ran to completion.
